Workout Review frame + start Workout Edit frame

+++ app/src/main/java/com/marcomadalin/olympus/data/database/entities/SetEntity.kt
@@ -10,10 +10,10 @@ data class SetEntity(
     @PrimaryKey(autoGenerate = true)
     var id : Long = 0,
     var exerciseId : Long = 0,
-    var weight : Int = 0,
+    var weight : Double = 0.0,
     var reps : Int = 0,
     var rir : Int = 0,
-    var lastWeight : Int = 0,
+    var lastWeight : Double = 0.0,
     var lastReps : Int = 0,
     var type : String = "",
     var setNumber : Int = 0,
