+++ app/src/main/java/com/marcomadalin/olympus/domain/model/Set.kt
@@ -6,10 +6,10 @@ import com.marcomadalin.olympus.domain.model.enums.SetType
 data class Set(
     var id : Long = 0,
     var exerciseId : Long = 0,
-    var weight : Int = 0,
+    var weight : Double = 0.0,
     var reps : Int = 0,
     var rir : Int = 0,
-    var lastWeight : Int = 0,
+    var lastWeight : Double = 0.0,
     var lastReps : Int = 0,
     var type : SetType = SetType.Warmup,
     var setNumber : Int = 0,
